+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## Upcoming
 
-- added internal binary crate `pp-gen` to calculate difficulty & pp values via `PerformanceCalculator.dll`
+- [BREAKING] The crate features for osu!standard have been adjusted:
+  - `no_leniency` has been removed
+  - `no_sliders_no_leniency` has been renamed to `osu_fast`
+  - `all_included` has been renamed to `osu_precise` and now serves as default
+  Additionally, instead of importing through `rosu_pp::osu::{version}`, you now have to import through `rosu_pp::osu`
 - [BREAKING] Instead of returning `PpResult`, performance calculations now return `PerformanceAttributes` depending on the mode.
 - [BREAKING] Instead of returning `StarResult`, difficulty calculations now return `DifficultyAttributes` depending on the mode.
+- added internal binary crate `pp-gen` to calculate difficulty & pp values via `PerformanceCalculator.dll`
 - osu: Updated up to commit [6944151486e677bfd11f2390163aca9161defbbf](https://github.com/ppy/osu/commit/6944151486e677bfd11f2390163aca9161defbbf) (2021-10-27)
 
 # v0.2.3